resume link, centralize content, site link
