+++ src/components/content.js
@@ -1,3 +1,8 @@
+export const resume =
+  "https://docs.google.com/document/d/1Rr-T2pQ-8Xd39VyT3fBhqCLBSFoJIZXOBiuZuCwxD6E/preview";
+
+export const siteLink = "https://github.com/smhaley/smhaley_portfolio";
+
 export const pepContent = {
   heading: "Pretty Easy Privacy",
   subHeading: "Simplifying PGP encryption",
@@ -17,4 +22,14 @@ export const tabContent = {
   appLink: "https://smhaley.github.io/react-text-annotate-blend/",
   dlLink: "https://www.npmjs.com/package/react-text-annotate-blend",
   stack: ["React", "Typescript"],
+};
+
+export const aboutContent = {
+  heading: "Educator turned statistician turned developer",
+  overview:
+    "Back end, front end, data engineering, machine learning. I'm a passionate developer in love with problem solving. My present interest is web and app development, but I'm always open to new problems.",
+  current:
+    "Currently I am a React developer and data engineer with Travelers Insurance. My focus is developing apps for consuming and training Natural Language models.",
+  github: "https://github.com/smhaley/",
+  linkedIn: "https://www.linkedin.com/in/shawn-haley/",
 };
