+++ src/components/nav/Menu.js
@@ -1,11 +1,15 @@
 import React, { useLayoutEffect } from "react";
 import { StyledMenu, MenuLink, MenuBtn } from "./styled/Menu.styled";
 import { bool } from "prop-types";
-
+import { resume } from "../content";
 const Menu = ({ open, handleMenuClick }) => {
   const options = ["About", "Work", "Contact"];
 
-  const navButton = <MenuBtn>Resume</MenuBtn>;
+  const navButton = (
+    <a href={resume} target="_blank" rel="noreferrer">
+      <MenuBtn>Resume</MenuBtn>
+    </a>
+  );
 
   useLayoutEffect(() => {
     if (open) {
