+++ src/components/project/ProjectContainer.js
@@ -3,7 +3,7 @@ import { CSSTransition } from "react-transition-group";
 import pep from "../../pics/pep.png";
 import tab from "../../pics/tab.png";
 import Project from "./Project";
-import { pepContent, tabContent } from "./content";
+import { pepContent, tabContent } from "../content";
 import { Heading } from "../../styles/common/generic";
 import { ProjContainer, Stack } from "./Project.styled";
